+++ fabric/src/main/java/dev/wuffs/playerplates/data/DataGens.java
@@ -1,5 +1,9 @@
 package dev.wuffs.playerplates.data;
 
+import dev.wuffs.playerplates.data.lang.ENUSLangGenerator;
+import dev.wuffs.playerplates.data.lang.PLPLLangGenerator;
+import dev.wuffs.playerplates.data.lang.ZHCNLangGenerator;
+import dev.wuffs.playerplates.data.lang.ZHTWLangGenerator;
 import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
 
@@ -8,5 +12,12 @@ public class DataGens implements DataGeneratorEntrypoint {
     public void onInitializeDataGenerator(FabricDataGenerator fabricDataGenerator) {
         FabricDataGenerator.Pack pack = fabricDataGenerator.createPack();
         pack.addProvider(RecipeGenerator::new);
+        pack.addProvider(ModelGenerator::new);
+
+        // Lang Gen
+        pack.addProvider(ENUSLangGenerator::new);
+        pack.addProvider(PLPLLangGenerator::new);
+        pack.addProvider(ZHCNLangGenerator::new);
+        pack.addProvider(ZHTWLangGenerator::new);
     }
 }